+++ src/math_c/vector_2.c
@@ -12,20 +12,20 @@
 
 #include "minirt_math.h"
 
-double magnitude(t_vec4d v)
+float magnitude(t_vec4d v)
 {
 	return (sqrt(v.x * v.x + v.y * v.y + v.z * v.z));
 }
 
 t_vec4d normalize(t_vec4d v)
 {
-	double mag_v;
+	float mag_v;
 
 	mag_v = magnitude(v);
 	return (create_vec4d(v.x / mag_v, v.y / mag_v, v.z / mag_v));
 }
 
-double dot(t_vec4d v1, t_vec4d v2)
+float dot(t_vec4d v1, t_vec4d v2)
 {
 	return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z + v1.p * v2.p);
 }
